v1

+++ frontend/src/util/formatDate.tsx
@@ -1,5 +1,5 @@
 export default function formatDate(dateString: string): string {
     const date: Date = new Date(dateString);
-    const options: Intl.DateTimeFormatOptions = { day: 'numeric', month: 'short', year: 'numeric' };
+    const options: Intl.DateTimeFormatOptions = { day: 'numeric', month: 'long', year: 'numeric' };
     return date.toLocaleDateString('en-EN', options);
 }
